Update all price and fee to be float type.

+++ db/schema.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS lifull_house_link (
 	house_id VARCHAR(32) NOT NULL PRIMARY KEY,
 	is_pr_item BOOLEAN NOT NULL DEFAULT FALSE,
 	listing_house_name VARCHAR(255) NOT NULL,
-	listing_house_price INT NOT NULL,
+	listing_house_price FLOAT NOT NULL,
 	sale_category VARCHAR(255) NOT NULL,
 	city VARCHAR(255) NOT NULL,
 	is_available BOOLEAN NOT NULL,
@@ -16,8 +16,8 @@ CREATE TABLE IF NOT EXISTS lifull_rent_link (
 	house_id VARCHAR(64) NOT NULL PRIMARY KEY,
 	is_pr_item BOOLEAN NOT NULL DEFAULT FALSE,
 	listing_house_name VARCHAR(255) NOT NULL,
-	listing_house_rent INT NOT NULL,
-	listing_house_manage_fee INT NOT NULL,
+	listing_house_rent FLOAT NOT NULL,
+	listing_house_manage_fee FLOAT NOT NULL,
 	city VARCHAR(255) NOT NULL,
 	is_available BOOLEAN NOT NULL,
 	first_available_date DATE NOT NULL,
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS lifull_rent_link (
 
 CREATE TABLE IF NOT EXISTS lifull_house_price_history (
 	house_id VARCHAR(32) NOT NULL,
-	price INT NOT NULL,
+	price FLOAT NOT NULL,
 	price_date DATE NOT NULL,
 	PRIMARY KEY(house_id, price_date)
 );
@@ -47,10 +47,10 @@ CREATE TABLE IF NOT EXISTS lifull_stations_near_house (
 CREATE TABLE IF NOT EXISTS lifull_house_info (
 	house_id VARCHAR(32) NOT NULL PRIMARY KEY,
 	name VARCHAR(255) NOT NULL,
-	price INT NOT NULL,
+	price FLOAT NOT NULL,
 	address VARCHAR(255) NOT NULL,
-	moneykyoueki INT NOT NULL,
-	moneyshuuzen INT NOT NULL,
+	moneykyoueki FLOAT NOT NULL,
+	moneyshuuzen FLOAT NOT NULL,
 	district VARCHAR(255) NOT NULL,
 	build_date DATE NOT NULL,
 	room VARCHAR(255),
@@ -72,11 +72,11 @@ CREATE TABLE IF NOT EXISTS lifull_house_info (
 	land_usage VARCHAR(255),
 	land_position VARCHAR(255),
 	land_right VARCHAR(32),
-	land_moneyshakuchi INT,
+	land_moneyshakuchi FLOAT,
 	land_term VARCHAR(255),
 	land_landkokudoho VARCHAR(32),
 	other_fee_details TEXT,
-	total_other_fee INT NOT NULL,
+	total_other_fee FLOAT NOT NULL,
 	manage_details TEXT,
 	latest_rent_status VARCHAR(32),
 	trade_method VARCHAR(32)
